Split out convolution_wrapper
